CUETools.Converter: can specify encoders and decoders in command line now CUETools.Converter: support for "nul" output (decode/test only)

+++ CUETools.Codecs/WAVWriter.cs
@@ -174,7 +174,8 @@ namespace CUETools.Codecs
             _bw.Close();
             _bw = null;
             _IO = null;
-            File.Delete(_path);
+            if (_path != "")
+                File.Delete(_path);
         }
 
         public void Write(AudioBuffer buff)
